feat: 포인트 결제 기능 구현

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/user/UserV1Dto.java
@@ -29,7 +29,7 @@ public class UserV1Dto {
 		}
 	}
 
-	public record UserPointResponse(int pointValue) {
+	public record UserPointResponse(long pointValue) {
 		public static UserPointResponse from(PointInfo pointInfo) {
 			return new UserPointResponse(pointInfo.pointValue());
 		}
